fix(artifacts): deleted unnecessary schema (patchArtifactSchema.ts)
